+++ app/src/main/java/com/goshoppi/pos/architecture/workmanager/SyncWorker.kt
@@ -52,7 +52,7 @@ class SyncWorker(private var context: Context, params: WorkerParameters) : Worke
 
                         response.body()?.data?.products!!.forEach {
                             it.variants.forEach {variant ->
-                                variant.productId = it.storeProductId
+                                variant.productId = it.storeProductId.toLong()
                                 masterVariantRepository.insertMasterVariant(variant)
                             }
                         }
